fix: extract matchesAtBoundary to shared nav utility
Extract path-matching helpers (matchesAtBoundary, findActiveApp,
isPageActive) from PageNav to app/nav/path-utils.ts. Update AppRail
and Sidebar to use boundary-safe matching, fixing prefix-collision
bugs where /fin could match /finance.
