+++ apps/pops-shell/src/app/layout/PageNav.tsx
@@ -8,35 +8,7 @@
 import { Link, useLocation } from "react-router";
 import { registeredApps } from "@/app/nav/registry";
 import { iconMap } from "@/app/nav/icon-map";
-import type { AppNavConfig } from "@/app/nav/types";
-
-/** Check if pathname matches a prefix at a path-segment boundary. */
-function matchesAtBoundary(pathname: string, prefix: string): boolean {
-  if (!pathname.startsWith(prefix)) return false;
-  // Must match exactly or be followed by / (segment boundary)
-  return pathname.length === prefix.length || pathname[prefix.length] === "/";
-}
-
-/** Find the active app by matching the current pathname against registered base paths. */
-export function findActiveApp(
-  pathname: string,
-  apps: AppNavConfig[],
-): AppNavConfig | undefined {
-  return apps.find((app) => matchesAtBoundary(pathname, app.basePath));
-}
-
-/** Check if a page item is active given the current pathname and its app's basePath. */
-export function isPageActive(
-  pathname: string,
-  basePath: string,
-  itemPath: string,
-): boolean {
-  if (itemPath === "") {
-    return pathname === basePath || pathname === `${basePath}/`;
-  }
-  const fullPath = `${basePath}${itemPath}`;
-  return matchesAtBoundary(pathname, fullPath);
-}
+import { findActiveApp, isPageActive } from "@/app/nav/path-utils";
 
 export function PageNav() {
   const location = useLocation();
